refactor: reorganize game modules and remove unused components
- Moved `CompareScores` and `ProcessRollType` to `GameStep2_inGame` directory.
- Deleted `PlayerSignIn.d.ts` and `StartGame.ts` as they are no longer needed.
