--- src/Modules/PlayerSignIn/PlayerSignIn.d.ts
@@ -1,1 +1,0 @@
-export interface IPlayerSignInProps {}
